backend/GradubtBackend/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
defusedxml==0.8.0rc2
Django==4.2.7
django-cors-headers==4.3.1
django-filter==23.4
django-rest-framework==0.1.0
django-templated-mail==1.1.1
django-unfold==0.15.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.1
idna==3.4
oauthlib==3.2.2
Pillow==10.1.0
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
slugify==0.0.1
social-auth-app-django==5.4.0
social-auth-core==4.5.0
sqlparse==0.4.4
urllib3==2.0.7
whitenoise==6.6.0
